requirements.txt
langchain
langgraph
langchain-groq
python-dotenv
pinecone
pdfreader
fastapi
uvicorn
psycopg2
langchain-unstructured
langchain-community
langchain_huggingface
langchain-pinecone
sentence-transformers
python-multipart